TST: collect string arithmetic tests
