+++ pandas/tests/io/formats/test_format.py
@@ -1922,37 +1922,6 @@ class TestGenericArrayFormatter:
         assert res[0] == " [[True, True], [False, False]]"
         assert res[1] == " [[False, True], [True, False]]"
 
-    def test_2d_extension_type(self):
-        # GH 33770
-
-        # Define a stub extension type with just enough code to run Series.__repr__()
-        class DtypeStub(pd.api.extensions.ExtensionDtype):
-            @property
-            def type(self):
-                return np.ndarray
-
-            @property
-            def name(self):
-                return "DtypeStub"
-
-        class ExtTypeStub(pd.api.extensions.ExtensionArray):
-            def __len__(self) -> int:
-                return 2
-
-            def __getitem__(self, ix):
-                return [ix == 1, ix == 0]
-
-            @property
-            def dtype(self):
-                return DtypeStub()
-
-        series = Series(ExtTypeStub(), copy=False)
-        res = repr(series)  # This line crashed before #33770 was fixed.
-        expected = "\n".join(
-            ["0    [False True]", "1    [True False]", "dtype: DtypeStub"]
-        )
-        assert res == expected
-
 
 def _three_digit_exp():
     return f"{1.7e8:.4g}" == "1.7e+008"
